requirements.txt
langchain==0.3.*
langchain-community
langchain-huggingface
langgraph
openai

fastapi
uvicorn

pydantic
requests

pandas

pytest

loguru

python-dotenv

